requirements.txt
discord.py==1.7.3
google-auth-httplib2
google_api_python_client==2.40.0
google_auth_oauthlib==0.5.0
protobuf==3.19.4
python-dotenv==0.19.2
requests==2.22.0
uwuify==1.1.0
boto==2.49.0